feat(nce): add --refresh flag wiring upstream fetch into user cache
