+++ crates/riri-node-lifecycle/tests/data/expected.json
