+++ crates/xtask/Cargo.toml
@@ -10,11 +10,8 @@ publish = false
 anyhow = { workspace = true }
 chrono = { workspace = true }
 clap = { workspace = true }
-riri-node-lifecycle = { workspace = true }
-semver = { workspace = true, features = ["serde"] }
-serde = { workspace = true }
+riri-node-lifecycle = { workspace = true, features = ["refresh"] }
 serde_json = { workspace = true }
-ureq = { workspace = true }
 
 [lib]
 path = "src/lib.rs"
