+++ crates/xtask/src/lib.rs
@@ -3,4 +3,3 @@
 #![allow(clippy::missing_errors_doc, clippy::missing_panics_doc)]
 
 pub mod refresh_node;
-pub mod remote;
